Backend/qrdentalproject/requirements.txt
asgiref==3.7.2
Django==5.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework==3.15.1
pyqrcode==1.2.1
pypng
